refactor: move activity log grouping into signals
